+++ barista-http/src/main/java/com/ibm/runtimes/events/coffeeshop/health/BaristaReadinessCheck.java
@@ -1,0 +1,23 @@
+package com.ibm.runtimes.events.coffeeshop.health;
+
+import javax.enterprise.context.ApplicationScoped;
+
+import org.eclipse.microprofile.health.HealthCheck;
+import org.eclipse.microprofile.health.HealthCheckResponse;
+import org.eclipse.microprofile.health.Readiness;
+
+@Readiness
+@ApplicationScoped
+public class BaristaReadinessCheck implements HealthCheck {
+
+    private boolean isReady() {
+        return true;
+    }
+	
+    @Override
+    public HealthCheckResponse call() {
+        boolean up = isReady();
+        return HealthCheckResponse.named(this.getClass().getSimpleName()).state(up).build();
+    }
+    
+}
