+++ barista-kafka/src/main/java/com/ibm/runtimes/events/coffeeshop/KafkaBarista.java
@@ -1,0 +1,55 @@
+package com.ibm.runtimes.events.coffeeshop;
+
+import org.eclipse.microprofile.config.inject.ConfigProperty;
+import org.eclipse.microprofile.reactive.messaging.Incoming;
+import org.eclipse.microprofile.reactive.messaging.Outgoing;
+
+import javax.enterprise.context.ApplicationScoped;
+import javax.json.bind.Jsonb;
+import javax.json.bind.JsonbBuilder;
+import javax.inject.Inject;
+
+import java.util.Random;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.CompletionStage;
+import java.util.concurrent.Executor;
+import java.util.concurrent.Executors;
+
+@ApplicationScoped
+public class KafkaBarista {
+
+    @Inject
+    @ConfigProperty(name = "barista.name")
+    String name;
+
+    private Random random = new Random();
+    
+    @Incoming("orders")
+    @Outgoing("queue")
+    public CompletionStage<String> prepare(String message) {
+        Jsonb jsonb = JsonbBuilder.create();
+        Order order = jsonb.fromJson(message, Order.class);
+        System.out.println("Barista " + name + " is going to prepare a " + order.getProduct());
+        return makeIt(order)
+                .thenApply(beverage -> PreparationState.ready(order, beverage));
+    }
+
+    private CompletionStage<Beverage> makeIt(Order order) {
+        return CompletableFuture.supplyAsync(() -> {
+            System.out.println("Preparing a " + order.getProduct());
+            prepare();
+            return new Beverage(order, name);
+        }, executor);
+    }
+
+    private Executor executor = Executors.newSingleThreadExecutor();
+
+    private void prepare() {
+        try {
+            Thread.sleep(random.nextInt(5000));
+        } catch (InterruptedException e) {
+            Thread.currentThread().interrupt();
+        }
+    }
+
+}
