+++ barista-kafka/src/main/java/com/ibm/runtimes/events/coffeeshop/health/BaristaLivenessCheck.java
@@ -1,0 +1,23 @@
+package com.ibm.runtimes.events.coffeeshop.health;
+
+import javax.enterprise.context.ApplicationScoped;
+
+import org.eclipse.microprofile.health.HealthCheck;
+import org.eclipse.microprofile.health.HealthCheckResponse;
+import org.eclipse.microprofile.health.Liveness;
+
+@Liveness
+@ApplicationScoped
+public class BaristaLivenessCheck implements HealthCheck {
+
+    private boolean isAlive() {
+        return true;
+    }
+	
+    @Override
+    public HealthCheckResponse call() {
+        boolean up = isAlive();
+        return HealthCheckResponse.named(this.getClass().getSimpleName()).state(up).build();
+    }
+    
+}
