--- barista-kafka/src/main/java/me/escoffier/quarkus/coffeeshop/health/BaristaLivenessCheck.java
@@ -1,23 +1,0 @@
-package me.escoffier.quarkus.coffeeshop.health;
-
-import javax.enterprise.context.ApplicationScoped;
-
-import org.eclipse.microprofile.health.HealthCheck;
-import org.eclipse.microprofile.health.HealthCheckResponse;
-import org.eclipse.microprofile.health.Liveness;
-
-@Liveness
-@ApplicationScoped
-public class BaristaLivenessCheck implements HealthCheck {
-
-    private boolean isAlive() {
-        return true;
-    }
-	
-    @Override
-    public HealthCheckResponse call() {
-        boolean up = isAlive();
-        return HealthCheckResponse.named(this.getClass().getSimpleName()).state(up).build();
-    }
-    
-}
