+++ coffeeshop-service/src/main/java/com/ibm/runtimes/events/coffeeshop/CoffeeShopApplication.java
@@ -1,0 +1,9 @@
+package com.ibm.runtimes.events.coffeeshop;
+
+import javax.ws.rs.ApplicationPath;
+import javax.ws.rs.core.Application;
+
+@ApplicationPath("/services")
+public class CoffeeShopApplication extends Application {
+
+}
