+++ coffeeshop-service/src/main/java/com/ibm/runtimes/events/coffeeshop/model/JsonProvider.java
@@ -1,0 +1,15 @@
+package com.ibm.runtimes.events.coffeeshop.model;
+
+import javax.enterprise.context.ApplicationScoped;
+import javax.enterprise.inject.Produces;
+import javax.json.bind.Jsonb;
+import javax.json.bind.JsonbBuilder;
+
+@ApplicationScoped
+public class JsonProvider {
+
+    @Produces
+    Jsonb json() {
+        return JsonbBuilder.create();
+    }
+}
