+++ coffeeshop-service/src/main/java/com/ibm/runtimes/events/coffeeshop/model/Order.java
@@ -1,0 +1,35 @@
+package com.ibm.runtimes.events.coffeeshop.model;
+
+public class Order {
+
+    private String product;
+    private String name;
+    private String orderId;
+
+    public String getProduct() {
+        return product;
+    }
+
+    public Order setProduct(String product) {
+        this.product = product;
+        return this;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public Order setName(String name) {
+        this.name = name;
+        return this;
+    }
+
+    public String getOrderId() {
+        return orderId;
+    }
+
+    public Order setOrderId(String orderId) {
+        this.orderId = orderId;
+        return this;
+    }
+}
