--- coffeeshop-service/src/main/java/me/escoffier/quarkus/coffeeshop/CoffeeShopApplication.java
@@ -1,9 +1,0 @@
-package me.escoffier.quarkus.coffeeshop;
-
-import javax.ws.rs.ApplicationPath;
-import javax.ws.rs.core.Application;
-
-@ApplicationPath("/services")
-public class CoffeeShopApplication extends Application {
-
-}
